Auto-fix yaml rules
In order to reduce divergance with ansible-lint rules, we apply
auto-fixing of violations.

In current patch we replace all kind of truthy variables with
`true` or `false` values to align with recommendations along with
alignment of used quotes.

With ansible-lint update it seems to be resolving symlink to absolute
path, which causes our linter check to search for the wrong
role name. Some more details were posted to ansible forum[1]

So we apply change here to resolve failing CI.

[1] https://forum.ansible.com/t/ansible-lint-resolving-symlinks-to-their-destinations/40388

Change-Id: Icdbe2bb4ad139254266d456cebba0c3f0c82496e